Refactor method declarations
