+++ sdk/sui/encoder.go
@@ -1,9 +1,11 @@
 package sui
 
 import (
+	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"math/big"
+	"strings"
 
 	"github.com/aptos-labs/aptos-go-sdk/bcs"
 	"github.com/block-vision/sui-go-sdk/models"
@@ -206,5 +208,26 @@ func serializeTimelockCancel(id []byte) ([]byte, error) {
 	return bcs.SerializeSingle(func(ser *bcs.Serializer) {
 		// Serialize the operation ID
 		ser.WriteBytes(id)
+	})
+}
+
+// serializeAuthorizeUpgradeParams serializes parameters for `mcms_deployer::authorize_upgrade`
+func serializeAuthorizeUpgradeParams(policy uint8, digest []byte, packageAddress string) ([]byte, error) {
+	// The authorize_upgrade function expects:
+	// - policy: u8
+	// - digest: vector<u8>
+	// - package_address: address
+
+	// Convert package address to bytes for BCS serialization
+	packageAddrBytes, err := hex.DecodeString(strings.TrimPrefix(packageAddress, "0x"))
+	if err != nil {
+		return nil, fmt.Errorf("failed to decode package address: %w", err)
+	}
+
+	// Use proper BCS serialization like the existing codebase
+	return bcs.SerializeSingle(func(ser *bcs.Serializer) {
+		ser.U8(policy)                   // u8 policy
+		ser.WriteBytes(digest)           // vector<u8> digest
+		ser.FixedBytes(packageAddrBytes) // address (32-byte fixed bytes)
 	})
 }
