+++ sdk/sui/transaction.go
@@ -1,6 +1,7 @@
 package sui
 
 import (
+	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -8,6 +9,8 @@ import (
 	"github.com/block-vision/sui-go-sdk/models"
 
 	"github.com/smartcontractkit/mcms/types"
+
+	"github.com/smartcontractkit/chainlink-sui/bindings/bind"
 )
 
 func ValidateAdditionalFields(additionalFields json.RawMessage) error {
@@ -89,4 +92,43 @@ func NewTransactionWithUpgradeData(moduleName, function string, to string, data 
 		Data:             data,
 		AdditionalFields: marshalledAdditionalFields,
 	}, nil
+}
+
+// CreateUpgradeTransaction creates a transaction for upgrading a package through MCMS
+func CreateUpgradeTransaction(compiledPackage bind.PackageArtifact, mcmsPackageID, depStateObj, registryObj, mcmsUserPackageId string) (types.Transaction, error) {
+	upgradePolicy := uint8(0) // Compatible upgrade policy
+	data, err := serializeAuthorizeUpgradeParams(upgradePolicy, compiledPackage.Digest, mcmsUserPackageId)
+	if err != nil {
+		return types.Transaction{}, fmt.Errorf("serializing authorize upgrade params: %w", err)
+	}
+
+	// Convert modules from base64 strings to raw bytes
+	moduleBytes := make([][]byte, len(compiledPackage.Modules))
+	for i, moduleBase64 := range compiledPackage.Modules {
+		decoded, err := base64.StdEncoding.DecodeString(moduleBase64)
+		if err != nil {
+			return types.Transaction{}, fmt.Errorf("decoding module %d: %w", i, err)
+		}
+		moduleBytes[i] = decoded
+	}
+
+	depAddresses := make([]models.SuiAddress, len(compiledPackage.Dependencies))
+	for i, dep := range compiledPackage.Dependencies {
+		depAddresses[i] = models.SuiAddress(dep)
+	}
+
+	// Create transaction targeting mcms_deployer::authorize_upgrade with upgrade data
+	return NewTransactionWithUpgradeData(
+		"mcms_deployer",             // Module name
+		"authorize_upgrade",         // Function name
+		mcmsPackageID,               // Package ID (mcms_deployer is in MCMS package)
+		data,                        // BCS-encoded parameters
+		"MCMS",                      // Contract type
+		[]string{"upgrade", "mcms"}, // Tags
+		depStateObj,                 // Main state object (DeployerState)
+		[]string{registryObj},       // Internal objects (Registry for validation)
+		moduleBytes,                 // Compiled modules for upgrade
+		depAddresses,                // Dependencies for upgrade
+		mcmsUserPackageId,           // Package being upgraded
+	)
 }
